fix: デプロイエラー解消のためworkoutページ内の関数を移行
